Cruster 0.4.5 - ui: execution of requests in repeater moved to additional module
